uart: refactor uart module to build async/await serial structure
The original serial structures were renamed to `BlockingSerial`, `BlockingTransmitHalf` and `BlockingReceiveHalf`. The new `Serial` structure dedicates to async/await function.

Serial constructors now returns error instead of panicking in the HAL crate. Examples are modified to unwrap the return value of the constructor. Ref: https://docs.embassy.dev/embassy-stm32/git/stm32f746ve/usart/struct.Uart.html#method.new

The serial `uart` module is rearranged into files for clearance of code.
