+++ bouffalo-hal/Cargo.toml
@@ -20,6 +20,8 @@ as-slice = "0.2.1"
 nb = "1.1.0"
 # For backward compatibility only. 
 embedded-hal-027 = { package = "embedded-hal", version = "0.2.7" }
+embedded-io-async = "0.6.1"
+atomic-waker = "1.1.2"
 
 [dev-dependencies]
 memoffset = "0.9.0"
