+++ bouffalo-hal/src/uart/error.rs
@@ -1,0 +1,32 @@
+/// Serial error.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum Error {
+    /// Framing error.
+    Framing,
+    /// Noise error.
+    Noise,
+    /// RX buffer overrun.
+    Overrun,
+    /// Parity check error.
+    Parity,
+}
+
+impl embedded_io::Error for Error {
+    #[inline(always)]
+    fn kind(&self) -> embedded_io::ErrorKind {
+        embedded_io::ErrorKind::Other
+    }
+}
+
+impl embedded_hal_nb::serial::Error for Error {
+    #[inline(always)]
+    fn kind(&self) -> embedded_hal_nb::serial::ErrorKind {
+        match self {
+            Error::Framing => embedded_hal_nb::serial::ErrorKind::FrameFormat,
+            Error::Noise => embedded_hal_nb::serial::ErrorKind::Noise,
+            Error::Overrun => embedded_hal_nb::serial::ErrorKind::Overrun,
+            Error::Parity => embedded_hal_nb::serial::ErrorKind::Parity,
+        }
+    }
+}
