+++ examples/peripherals/uart-async-demo/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "uart-async-demo"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bouffalo-hal = { path = "../../../bouffalo-hal", features = ["bl808"] }
+bouffalo-rt = { path = "../../../bouffalo-rt", features = ["bl808-dsp"] }
+panic-halt = "0.2.0"
+embedded-time = "0.12.1"
+riscv = "0.11.1"
+
+[[bin]]
+name = "uart-async-demo"
+test = false
