+++ examples/peripherals/uart-async-demo/README.md
@@ -1,0 +1,8 @@
+Async/await UART peripheral demo
+
+Build this example with:
+
+```
+rustup target install riscv64imac-unknown-none-elf
+cargo build --target riscv64imac-unknown-none-elf --release -p uart-async-demo
+```
